Create README.md
For the creation of the board of the game "Serpents and ladders" the code taught by Professor Juan Manuel Reyes was used. The explanatory videos found in the instructions for integrative task 2.
This project is about the famous game "snakes and ladders" brought to a console graphical interface. For the development of this task, topics such as linked lists, doubly linked, recursive lists and some binary tree concepts were used.